use positron api npm package and rename accordingly
